requirements.txt
anthropic>=0.43.0
exa-py>=1.8.0
tldextract>=5.1.0
jinja2>=3.1.0
feedparser>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0
